Fixed an issue with exposing the millimeter Pint unit.
The Python startup has been updated to fix an issue with Pint's
millimeter unit not being exposed. Some of the SI unit prefixes of base
units aren't listed in the unit registry's unit list initially. Thus,
they must be accessed through the `__getattr__` method to be added to
the unit list.