Updated the Bash PS1 prompt to handle slow filesystems.
The PS1 prompt has been updated to call a new wrapper function
`get-git-ps1`. This function checks if the current working directory is
on a slow filesystem (e.g. network mounted drive), typically a FUSE. If
it is, then the Git portion of the prompt is skipped to prevent long and
slow delays between entering commands on the prompt.