Added a Git alias for renaming Git tags.
Since the process of renaming a Git tag is not straight forward, an
alias has been added to perform this, which reads the contents of the
old tag and then generates a new tag with the different name. The new
tag points to the same commit as the old one.