Moved complex Git aliases into proper Bash scripts.
Previously, the more complex Git aliases (which required a function in
order to parse some arguments) were coded inside the Git configuration
file. However, due to the nature of how the command is invoked, and the
general fact that coding Bash outside its native environment is messy,
these were messy and not the best to maintain.

Instead, these Git aliases were moved to be implemented as scripts. When
running a subcommand, Git will look for any script in the user's path
that matches the format `git-<subcommand_name` (e.g. `git-ls-authors`).
Thus, a new `git_subcommands` directory was created and added to the
path in the Bashrc file. Additionally, the setup script was updated to
link to the new directory.