Added a notify function variant to send a push notification to Bash.
A new `notify-phone` function has been added to the Bashrc file. This
performs the same functionality as the other notify functions, except
that it sends a push notification to a phone when the command finishes.

This uses the If This Then That (IFTTT) service paired with a companion
app on a phone to send the push notification. The command sends a POST
request with Curl to a specific URL, which in turn triggers a push
notification with the specified message in the POST data. As this URL
can be used to send a bunch of notifications to a private phone, the
user is expected to encode the URL in an environment variable (ideally
in a private, local Bashrc file).