Extended the notify function for Bash to support no command.
This  allows the user to simply send a notification. This use case is
for if they simply want to send a notification, or send a notification
after another command (e.g. `command ; notify-phone "Command"`). This
can be useful for more complex commands such as Bash for loops which are
difficult to `eval` directly.