Adjusted how the no command case is handled for the notify functions.
The Bash notify functions now treat the first argument as the complete
message to display in the notification when no command is specified
(only a single argument is specified to the function). This gives the
user more flexibility when they are running this after a command and
also allows the function to just send a plain notification.