Added back support for running aliases to the notify functions.
During the changes to the Bash notify functions, support was
accidentally removed for running aliases. This is because the command
(first argument) was quoted. Quotes around a command (e.g. `"msbuild"`)
mean that Bash will treat it as a literal command and not try to expand
it as an alias first. Thus, this has been reverted.

However, the implication is that the command will no longer be able to
run commands with spaces in them. This is less important than running
aliases, however.