Added sourcing of the Broot configuration to the Bashrc file.
This configuration file simply sets up a wrapper function `br` for the
main `Broot` program. This is necessary because several actions in
Broot, such as changing the directory, will only be useful if they are
run in the current process, not in a subprocess.

Broot is a tree-based navigation tool for navigating large directory
trees, with support for various command and the ability to fuzzy find
paths.