Adding warning messages to Bashrc if expected files don't exist.
To set up appropriate defaults and other settings, the Bashrc first
sources system-wide and command-specific configuration files. These have
guards to ensure the files only get sourced if they exist. Now, a
warning message is printed if the files do not exist so that the user
will be aware of missing files.

Additionally, some non-existent and not useful files being sourced were
removed.