Added a Vim plugin for highlighting Shakespeare templates in Haskell.
Shakespeare templates are used to embed templates for code in other
languages (e.g. HTML, CSS, etc.) in Haskell code. Without this plugin,
the templates will be attempted to be highlighted as Haskell code, which
will lead to many highlights of illegal syntax. This plugin ensures
these sections are instead highlighted as Shakespeare templates.