Updated the Python startup file to conform to linting standards.
All of the linting issues with the file were resolved by adding type
annotations, fixing unused imports, etc. In addition, several changes
were made to conform to more standard styles now used.

The code for getting color codes for the terminal was also greatly
simplified by having it only handle true colors and not presetting the
values in a dictionary.