Added the ALE plugin to Vim.
The Asynchronous Lint Engine (ALE) plugin has been added to Vim. This
plugin enables live (asynchronous) linting for a wide variety of
languages to Vim and displaying of these errors using the sign gutter
and status line.

This plugin has been added to Vim along with some default settings for
Python and standard shortcuts for navigating between the linting
violations.