Added a PIP requirements file to track global Python packages.
A `requirements.txt` file has been added to the repository to keep track
of the Python packages that are used globally by the user. These
consistent of utility packages like `see` and linting packages like
`flake8` and `mypy`.

Tracking the files in a PIP requirements file means they do not have to
be hardcoded in the system setup script and they can also be more easily
updated since PIP does not support a general update all packages command
(specific packages must be supplied).