Changed the default `rsync` behavior and added a new alias.
The default `rsync` command now no longer uses the `--checksum` option
(codified via an alias). The problem with this option is that it
unconditionally computes the checksum on all input files before starting
to copy files to the remote, even if those files don't exist on the
remote. This takes a huge amount of time.

To still support this checksum behavior when desired, a new
`rsync-strict` alias has been added to run with the `--checksum` option.
This is useful when a user wants to be certain the files match on both
sides.