Updated references to the FD command to `fdfind`.
The Bash configuration files have been updated to replace references to
the FD command as `fd` to `fdfind`. Since the FD program is now
installed through official Ubuntu APT repositories instead of through
the Rust toolchain, the program is installed as `fdfind` instead of
`fd` as the `fdclone` package uses that name already.

Also, the formatting for the aliases and commands were cleaned up.